#278: Extract proposal generation functions into utility file
